+++ webui/react/src/components/Searches/Searches.settings.ts
@@ -1,21 +1,10 @@
 import * as t from 'io-ts';
 
 import { INIT_FORMSET } from 'components/FilterForm/components/FilterFormStore';
+import { RegularSelectionType, SelectionType } from 'types';
 
 import { defaultColumnWidths, defaultExperimentColumns } from './columns';
 
-const SelectAllType = t.type({
-  exclusions: t.array(t.number),
-  type: t.literal('ALL_EXCEPT'),
-});
-
-const RegularSelectionType = t.type({
-  selections: t.array(t.number),
-  type: t.literal('ONLY_IN'),
-});
-
-export const SelectionType = t.union([RegularSelectionType, SelectAllType]);
-export type SelectionType = t.TypeOf<typeof SelectionType>;
 export const DEFAULT_SELECTION: t.TypeOf<typeof RegularSelectionType> = {
   selections: [],
   type: 'ONLY_IN',
