+++ webui/react/src/pages/F_ExpList/F_ExperimentList.tsx
@@ -80,6 +80,7 @@ import {
   ProjectColumn,
   ProjectMetricsRange,
   RunState,
+  SelectionType as SelectionState,
 } from 'types';
 import handleError from 'utils/error';
 import { getProjectExperimentForExperimentItem } from 'utils/experiment';
@@ -100,7 +101,6 @@ import {
   defaultProjectSettings,
   ProjectSettings,
   ProjectUrlSettings,
-  SelectionType as SelectionState,
   settingsPathForProject,
 } from './F_ExperimentList.settings';
